Fixing warnings

+++ engine/src/engine/renderer/text_manager.cpp
@@ -59,7 +59,7 @@ void engine::text_manager::init()
 				texture,
 				glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 				glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-				face->glyph->advance.x
+				(uint32_t)face->glyph->advance.x
 			};
 			characters.insert(std::pair<GLchar, engine::character>(c, character));
 		}
